Added 5 levels

+++ Assets/Scripts/Level/Level.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class Level : MonoBehaviour
 {
     public Transform levelStartPoint;
-    public float bubbleCount = 3;
+    public int bubbleCount = 3;
     public GameObject transitionAnchor;
 
     // Start is called once before the first execution of Update after the MonoBehaviour is created
